% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{package}
\name{BCyto}
\alias{BCyto}
\title{Analyse and visualize Flow Cytometry data}
\description{
The BCyto R package is an open-source project that
provides an user-friendly, high-performance UI for Flow Cytometry analysis.
}
