% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{package}
\name{csbSTATS}
\alias{csbSTATS-package}
\alias{csbSTATS}
\title{Analyse and visualize biostatistics data}
\description{
The csbSTATS R package is an open-source project that provides a
user-friendly, interactive UI for exploratory data analysis,
visualization, and statistical testing.
}
\author{
\strong{Maintainer}: Caio Santos Bonilha \email{Caio.Bonilha@glasgow.ac.uk}

}
