% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runBCyto}
\alias{runBCyto}
\title{Runs the BCyto app}
\usage{
runBCyto()
}
\description{
runBCyto() is a wrapper function of
shinyApp(ui, server, onStart).
}
\details{
Use runBCyto() at the R console for initiating the BCyto Shiny app.
}
