% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runCsbSTATS}
\alias{runCsbSTATS}
\title{Run the csbSTATS application}
\usage{
runCsbSTATS()
}
\description{
runCsbSTATS() launches the csbSTATS Shiny application.
}
\details{
Use runCsbSTATS() at the R console to initiate the csbSTATS UI.
}
